@@ -2,12 +2,12 @@
 
   <Import Project="..\..\..\..\..\tests.props" />
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="System.Linq.Expressions" Version="4.3.0" />
-    <PackageReference Include="Utilities" Version="1.0.0-*" />
+    <PackageReference Include="Candoumbe.MiscUtilities" Version="0.1.0" />
   </ItemGroup>
 
   <ItemGroup>
