@@ -1,32 +1,39 @@
 ﻿using Bogus;
+
 using FluentAssertions;
+
 using FluentValidation;
-using FluentValidation.Results;
+
 using Identity.DataStores;
 using Identity.DTO;
 using Identity.Objects;
+
 using MedEasy.DAL.EFStore;
 using MedEasy.DAL.Interfaces;
 using MedEasy.IntegrationTests.Core;
+
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+
 using Moq;
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
-using System.Text;
 using System.Threading.Tasks;
+
 using Xunit;
 using Xunit.Abstractions;
 using Xunit.Categories;
+
 using ValidationResult = FluentValidation.Results.ValidationResult;
 
 namespace Identity.Validators.UnitTests
 {
     [UnitTest]
     [Feature("Accounts")]
-    public class NewAccountInfoValidatorTests : IDisposable, IClassFixture<SqliteDatabaseFixture>
+    public class NewAccountInfoValidatorTests : IAsyncDisposable, IClassFixture<SqliteDatabaseFixture>
     {
         private ITestOutputHelper _outputHelper;
         private IUnitOfWorkFactory _uowFactory;
@@ -36,7 +43,7 @@ namespace Identity.Validators.UnitTests
         public NewAccountInfoValidatorTests(ITestOutputHelper outputHelper, SqliteDatabaseFixture databaseFixture)
         {
             _outputHelper = outputHelper;
-            DbContextOptionsBuilder<IdentityContext> dbContextOptionsBuilder = new DbContextOptionsBuilder<IdentityContext>();
+            DbContextOptionsBuilder<IdentityContext> dbContextOptionsBuilder = new();
             dbContextOptionsBuilder.UseSqlite(databaseFixture.Connection);
             _uowFactory = new EFUnitOfWorkFactory<IdentityContext>(dbContextOptionsBuilder.Options, (options) =>
             {
@@ -45,11 +52,11 @@ namespace Identity.Validators.UnitTests
                 return context;
             });
             _loggerMock = new Mock<ILogger<NewAccountInfoValidator>>();
-            
+
             _sut = new NewAccountInfoValidator(_uowFactory, _loggerMock.Object);
         }
 
-        public async void Dispose()
+        public async ValueTask DisposeAsync()
         {
             using (IUnitOfWork uow =_uowFactory.NewUnitOfWork())
             {
